Second refactor step
