+++ chaM3Leon/batch/src/test/java/com/smartshaped/chameleon/batch/BatchLayerTestClass.java
@@ -1,7 +1,7 @@
 package com.smartshaped.chameleon.batch;
 
 import com.smartshaped.chameleon.batch.exception.BatchLayerException;
-import com.smartshaped.chameleon.common.exception.ConfigurationException;
+import com.smartshaped.chameleon.common.utils.exception.ConfigurationException;
 
 public class BatchLayerTestClass extends BatchLayer {
