+++ chaM3Leon/batch/src/test/java/com/smartshaped/chameleon/batch/BatchUpdaterTest.java
@@ -25,8 +25,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.smartshaped.chameleon.batch.exception.BatchUpdaterException;
 import com.smartshaped.chameleon.batch.utils.BatchConfigurationUtils;
-import com.smartshaped.chameleon.common.exception.CassandraException;
-import com.smartshaped.chameleon.common.exception.ConfigurationException;
+import com.smartshaped.chameleon.common.utils.exception.CassandraException;
+import com.smartshaped.chameleon.common.utils.exception.ConfigurationException;
 import com.smartshaped.chameleon.common.utils.CassandraUtils;
 import com.smartshaped.chameleon.common.utils.TableModel;
