+++ chaM3Leon/batch/src/test/java/com/smartshaped/chameleon/batch/PreprocessorTestClass.java
@@ -3,8 +3,8 @@ package com.smartshaped.chameleon.batch;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
 
-import com.smartshaped.chameleon.preprocessing.Preprocessor;
-import com.smartshaped.chameleon.preprocessing.exception.PreprocessorException;
+import com.smartshaped.chameleon.common.preprocessing.Preprocessor;
+import com.smartshaped.chameleon.common.preprocessing.exception.PreprocessorException;
 
 public class PreprocessorTestClass extends Preprocessor {
