+++ chaM3Leon/batch/src/test/resources/test-config.yml
@@ -27,171 +27,3 @@ batch:
       topic2:
         name: "topicTif"
         class: "com.smartshaped.chameleon.batch.PreprocessorTestClass"
-speed:
-  spark:
-    app:
-      name: "ml-application"
-    master: "local[*]"
-    yarn:
-      am:
-        cores: 1
-        memory: "512m"
-        memoryOverhead: "384m"
-    driver:
-      memory: "512m"
-      memoryOverhead: "384m"
-      cores: 1
-      maxResultSize: "1g"
-    executor:
-      memoryOverhead: "384m"
-      memory: ""
-      instances: 2
-      cores: 1
-    serializer: "org.apache.spark.serializer.KryoSerializer"
-    memory:
-      fraction: 0.7
-      storageFraction: 0.5
-      offHeap:
-        enabled: true
-        size: "1g"
-    sql:
-      parquet:
-        compression:
-          codec: "zstd"
-      shuffle:
-        partitions: 200
-    hadoop:
-      parquet:
-        page:
-          size:
-            row:
-              check:
-                min: 2
-                max: 10
-      dfs:
-        replication: 1
-  hdfs:
-    filter:
-      polygon: ""
-    manipulation:
-      splittingFactorX: "0"
-      splittingFactorY: "0"
-      rescalingFactorX: "0"
-      rescalingFactorY: "0"
-    readers:
-      default: ""
-      reader1:
-        class: "com.smartshaped.chameleon.ml.CustomReader"
-        path: "/tif"
-  cassandra:
-    model:
-      class: "com.smartshaped.chameleon.common.com.smartshaped.chameleon.batch.utils.TableModelExample"
-    node: "127.0.0.1"
-    port: "9042"
-    datacenter: "datacenter1"
-    keyspace:
-      name: "test"
-      replication_factor: 1
-  kafka:
-    broker: "kafka1:9092"
-    server: "kafkaservers"
-    intervalMs: "30000"
-    topics:
-      topic1:
-        name: "topic1"
-        class: null
-      topic2:
-        name: "topicTif"
-        class: "TifProcessor"
-  updater:
-    class: "com.smartshaped.chameleon.speed.SpeedUpdaterTestClass"
-ml:
-  hdfs:
-    readers:
-      default: ""
-      reader1:
-        class: "com.smartshaped.chameleon.ml.CustomReader"
-        path: "/tif"
-  cassandra:
-    model:
-      class: "com.smartshaped.chameleon.common.com.smartshaped.chameleon.batch.utils.TableModelExample"
-    node: "127.0.0.1"
-    port: "9042"
-    datacenter: "datacenter1"
-    keyspace:
-      name: "test"
-      replication_factor: 1
-  blackbox:
-    class: "com.smartshaped.chameleon.ml.blackbox.PythonBlackboxExample"
-    inputs: "input"
-    output: "output"
-    modelPath: "model_path"
-    folder: "/user/spark/blackbox"
-    python:
-      scriptPath: "script.py"
-      extraScripts: "extra_script.py"
-      libraries: "library1,library2"
-      requirementsPath: "requirements.txt"
-      extraArguments:
-        key1: "value1"
-        key2: "value2"
-        key3: "value3"
-  pipeline:
-    class: "com.smartshaped.chameleon.ml.PipelineExample"
-  modelSaver:
-    class: "com.smartshaped.chameleon.ml.ModelSaverExample"
-harvester:
-  RequestHandler: "default"
-  downloader:
-    DownloaderTwoExample:
-      params:
-        variables: "var1, var2"
-    DownloaderExample:
-      params:
-        start: "20100101"
-        end: "20100201"
-        community: "ag"
-        parameters: "par1,par2,par3"
-        header: "true"
-        time-standard: "lst"
-        site-elevation: "0"
-      url:
-        schema: "https"
-        authority: "super.auth"
-        path: "/api/some/path"
-      path:
-        hdfs-path: "/user/spark/Downloads/CopernicusMaps"
-  preprocessor:
-    TifRescalerPreprocessor:
-      params:
-        splittingFactorX: "64"
-        splittingFactorY: "64"
-        rescalingFactorX: "0.01"
-        rescalingFactorY: "0.01"
-    HarvesterExample:
-      params:
-        splittingFactorX: "64"
-        splittingFactorY: "64"
-        rescalingFactorX: "0.01"
-        rescalingFactorY: "0.01"
-  harvesters:
-    harvester1:
-      inputPath: ""
-      class: "com.smartshaped.chameleon.harvester.HarvesterExample"
-      downloader:
-        class: "com.smartshaped.chameleon.harvester.DownloaderExample"
-      transformer: "com.smartshaped.chameleon.harvester.TransformerExample"
-      preprocessor: "com.smartshaped.chameleon.harvester.PreprocessorExample"
-      saver:
-        class: "HarvesterSaver"
-        path: "/some/path"
-        checkpoint: ""
-    harvester2:
-test:
-  spark:
-    master: "local[*]"
-    deploy-mode: "client"
-    app-name: "speed"
-    interval-sec: 600
-  cassandra:
-    checkpoint: ""
