--- chaM3Leon/common/src/main/resources/framework-config.yml
@@ -1,3 +1,0 @@
-env: local
-
-include: ${env}-config.yml
