--- chaM3Leon/common/src/main/resources/typeMapping.yml
@@ -1,11 +1,0 @@
-typeMapping:
-  UUID: "uuid"
-  int: "int"
-  Integer: "int"
-  String: "text"
-  long: "bigint"
-  Long: "bigint"
-  boolean: "boolean"
-  Boolean: "boolean"
-  Float: "float"
-  Timestamp: "timestamp"
