+++ chaM3Leon/common/src/test/java/com/smartshaped/chameleon/common/utils/ConfigurationUtilsExample.java
@@ -1,6 +1,6 @@
 package com.smartshaped.chameleon.common.utils;
 
-import com.smartshaped.chameleon.common.exception.ConfigurationException;
+import com.smartshaped.chameleon.common.utils.exception.ConfigurationException;
 
 public class ConfigurationUtilsExample extends ConfigurationUtils {
