+++ chaM3Leon/common/src/test/java/com/smartshaped/chameleon/common/utils/TableModelTest.java
@@ -2,8 +2,8 @@ package com.smartshaped.chameleon.common.utils;
 
 import org.junit.jupiter.api.Test;
 
-import com.smartshaped.chameleon.common.exception.CassandraException;
-import com.smartshaped.chameleon.common.exception.ConfigurationException;
+import com.smartshaped.chameleon.common.utils.exception.CassandraException;
+import com.smartshaped.chameleon.common.utils.exception.ConfigurationException;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
