+++ chaM3Leon/harvester/src/test/java/com/smartshaped/chameleon/harvester/HarvesterTest.java
@@ -18,7 +18,7 @@ import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import com.smartshaped.chameleon.common.exception.ConfigurationException;
+import com.smartshaped.chameleon.common.utils.exception.ConfigurationException;
 import com.smartshaped.chameleon.harvester.exception.DownloaderException;
 import com.smartshaped.chameleon.harvester.exception.HarvesterException;
 import com.smartshaped.chameleon.harvester.request.Request;
