+++ chaM3Leon/harvester/src/test/java/com/smartshaped/chameleon/harvester/downloader/BinaryDownloaderTest.java
@@ -9,7 +9,7 @@ import static org.mockito.Mockito.mockConstruction;
 import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.when;
 
-import com.smartshaped.chameleon.common.exception.ConfigurationException;
+import com.smartshaped.chameleon.common.utils.exception.ConfigurationException;
 import com.smartshaped.chameleon.harvester.exception.DownloaderException;
 import com.smartshaped.chameleon.harvester.request.Request;
 import com.smartshaped.chameleon.harvester.utils.HarvesterConfigurationUtils;
