+++ chaM3Leon/harvester/src/test/java/com/smartshaped/chameleon/harvester/downloader/FTPDownloaderTestClass.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.apache.commons.net.ftp.FTPClient;
 
-import com.smartshaped.chameleon.common.exception.ConfigurationException;
+import com.smartshaped.chameleon.common.utils.exception.ConfigurationException;
 import com.smartshaped.chameleon.harvester.exception.DownloaderException;
 import com.smartshaped.chameleon.harvester.request.Request;
