+++ chaM3Leon/harvester/src/test/java/com/smartshaped/chameleon/harvester/utils/HarvesterConfigurationUtilsTest.java
@@ -17,8 +17,8 @@ import org.mockito.Mock;
 import org.mockito.MockedStatic;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import com.smartshaped.chameleon.common.exception.CassandraException;
-import com.smartshaped.chameleon.common.exception.ConfigurationException;
+import com.smartshaped.chameleon.common.utils.exception.CassandraException;
+import com.smartshaped.chameleon.common.utils.exception.ConfigurationException;
 import com.smartshaped.chameleon.common.utils.CassandraUtils;
 
 @ExtendWith(MockitoExtension.class)
