+++ chaM3Leon/ml/src/test/java/com/smartshaped/chameleon/ml/CustomMlLayer.java
@@ -1,7 +1,7 @@
 package com.smartshaped.chameleon.ml;
 
-import com.smartshaped.chameleon.common.exception.CassandraException;
-import com.smartshaped.chameleon.common.exception.ConfigurationException;
+import com.smartshaped.chameleon.common.utils.exception.CassandraException;
+import com.smartshaped.chameleon.common.utils.exception.ConfigurationException;
 import com.smartshaped.chameleon.ml.exception.MLLayerException;
 
 public class CustomMlLayer extends MLLayer {
