+++ chaM3Leon/ml/src/test/java/com/smartshaped/chameleon/ml/utils/MLConfigurationUtilsTest.java
@@ -6,7 +6,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.when;
 
-import com.smartshaped.chameleon.common.exception.ConfigurationException;
+import com.smartshaped.chameleon.common.utils.exception.ConfigurationException;
 import com.smartshaped.chameleon.common.utils.CassandraUtils;
 import java.util.Iterator;
 import org.apache.commons.configuration2.YAMLConfiguration;
