+++ chaM3Leon/speed/src/test/java/com/smartshaped/chameleon/speed/SpeedLayerTest.java
@@ -1,6 +1,6 @@
 package com.smartshaped.chameleon.speed;
 
-import com.smartshaped.chameleon.common.exception.ConfigurationException;
+import com.smartshaped.chameleon.common.utils.exception.ConfigurationException;
 import com.smartshaped.chameleon.common.utils.CassandraUtils;
 import com.smartshaped.chameleon.common.utils.KafkaConsumer;
 import com.smartshaped.chameleon.speed.exception.SpeedLayerException;
