+++ chaM3Leon/speed/src/test/java/com/smartshaped/chameleon/speed/SpeedUpdaterTestClass.java
@@ -1,7 +1,7 @@
 package com.smartshaped.chameleon.speed;
 
-import com.smartshaped.chameleon.common.exception.CassandraException;
-import com.smartshaped.chameleon.common.exception.ConfigurationException;
+import com.smartshaped.chameleon.common.utils.exception.CassandraException;
+import com.smartshaped.chameleon.common.utils.exception.ConfigurationException;
 import com.smartshaped.chameleon.speed.exception.SpeedUpdaterException;
 
 import org.apache.spark.sql.Dataset;
